@@ -1,10 +1,8 @@
-/*eslint-disable*/
-import React from "react";
+import React, {useEffect, useState} from "react";
 import { NavLink as NavLinkRRD, Link } from "react-router-dom";
-// nodejs library to set properties for components
 import { PropTypes } from "prop-types";
 
-// reactstrap components
+// Components
 import {
   Collapse,
   DropdownMenu,
@@ -27,37 +25,40 @@ import {
   Col
 } from "reactstrap";
 
-var ps;
 
 // Assets
-import iconProfile from "../../assets/img/profile/newuser.jpg"
+import iconProfile from "assets/img/profile/newuser.jpg"
 
-class Sidebar extends React.Component {
-  state = {
+const Sidebar = (props) => {
+  const [state, setState] = useState({
     collapseOpen: false
-  };
-  constructor(props) {
+  });
+  /* constructor(props) {
     super(props);
     this.activeRoute.bind(this);
-  }
+  } */
   // verifies if routeName is the one active (in browser input)
-  activeRoute(routeName) {
+  const activeRoute = (routeName) => {
     return this.props.location.pathname.indexOf(routeName) > -1 ? "active" : "";
   }
+  /* useEffect(()=>{
+    activeRoute()
+  },[]) */
   // toggles collapse between opened and closed (true/false)
-  toggleCollapse = () => {
-    this.setState({
-      collapseOpen: !this.state.collapseOpen
+  const toggleCollapse = () => {
+    setState({
+      collapseOpen: !state.collapseOpen
     });
   };
   // closes the collapse
-  closeCollapse = () => {
-    this.setState({
+  const closeCollapse = () => {
+    setState({
       collapseOpen: false
     });
   };
   // creates the links that appear in the left menu / Sidebar
-  createLinks = routes => {
+  const createLinks = routes => {
+    console.log(routes)
     return routes.map((prop, key) => {
       if (prop.sidebar == true){
         return (
@@ -65,7 +66,7 @@ class Sidebar extends React.Component {
             <NavLink
               to={prop.layout + prop.path}
               tag={NavLinkRRD}
-              onClick={this.closeCollapse}
+              onClick={closeCollapse}
               activeClassName="active"
             >
               <i className={prop.icon} />
@@ -76,12 +77,12 @@ class Sidebar extends React.Component {
       }
     });
   };
-  logout() {
+  const logout = () => {
     localStorage.clear();
     window.location = "/";
   }
-  render() {
-    const { bgColor, routes, logo } = this.props;
+
+    const { bgColor, routes, logo } = props;
     let navbarBrandProps;
     if (logo && logo.innerLink) {
       navbarBrandProps = {
@@ -105,7 +106,7 @@ class Sidebar extends React.Component {
           <button
             className="navbar-toggler"
             type="button"
-            onClick={this.toggleCollapse}
+            onClick={toggleCollapse}
           >
             <span className="navbar-toggler-icon" />
           </button>
@@ -155,20 +156,8 @@ class Sidebar extends React.Component {
                   <i className="ni ni-single-02" />
                   <span>My profile</span>
                 </DropdownItem>
-                {/*<DropdownItem to="/admin/user-profile" tag={Link}>
-                  <i className="ni ni-settings-gear-65" />
-                  <span>Settings</span>
-                </DropdownItem>
-                <DropdownItem to="/admin/user-profile" tag={Link}>
-                  <i className="ni ni-calendar-grid-58" />
-                  <span>Activity</span>
-                </DropdownItem>
-                <DropdownItem to="/admin/user-profile" tag={Link}>
-                  <i className="ni ni-support-16" />
-                  <span>Support</span>
-                </DropdownItem>*/}
                 <DropdownItem divider />
-                <DropdownItem onClick={this.logout}>
+                <DropdownItem onClick={logout}>
                     <i className="ni ni-user-run" />
                     <span>Logout</span>
                   </DropdownItem>
@@ -176,7 +165,7 @@ class Sidebar extends React.Component {
             </UncontrolledDropdown>
           </Nav>
           {/* Collapse */}
-          <Collapse navbar isOpen={this.state.collapseOpen}>
+          <Collapse navbar isOpen={state.collapseOpen}>
             {/* Collapse header */}
             <div className="navbar-collapse-header d-md-none">
               <Row>
@@ -197,7 +186,7 @@ class Sidebar extends React.Component {
                   <button
                     className="navbar-toggler"
                     type="button"
-                    onClick={this.toggleCollapse}
+                    onClick={toggleCollapse}
                   >
                     <span />
                     <span />
@@ -222,45 +211,15 @@ class Sidebar extends React.Component {
               </InputGroup>
             </Form>
             {/* Navigation */}
-            <Nav navbar>{this.createLinks(routes)}</Nav>
+            <Nav navbar>{createLinks(routes)}</Nav>
             {/* Divider */}
             <hr className="my-3" />
             {/* Heading */}
             <h6 className="navbar-heading text-muted">Documentation</h6>
-            {/* Navigation 
-            <Nav className="mb-md-3" navbar>
-              <NavItem>
-                <NavLink href="https://demos.creative-tim.com/argon-dashboard-react/#/documentation/overview?ref=adr-admin-sidebar">
-                  <i className="ni ni-spaceship" />
-                  Getting started
-                </NavLink>
-              </NavItem>
-              <NavItem>
-                <NavLink href="https://demos.creative-tim.com/argon-dashboard-react/#/documentation/colors?ref=adr-admin-sidebar">
-                  <i className="ni ni-palette" />
-                  Foundation
-                </NavLink>
-              </NavItem>
-              <NavItem>
-                <NavLink href="https://demos.creative-tim.com/argon-dashboard-react/#/documentation/alerts?ref=adr-admin-sidebar">
-                  <i className="ni ni-ui-04" />
-                  Components
-                </NavLink>
-              </NavItem>
-            </Nav>
-            <Nav className="mb-md-3" navbar>
-              <NavItem className="active-pro active">
-                <NavLink href="https://www.creative-tim.com/product/argon-dashboard-pro-react?ref=adr-admin-sidebar">
-                  <i className="ni ni-spaceship" />
-                  Upgrade to PRO
-                </NavLink>
-              </NavItem>
-            </Nav>*/}
           </Collapse>
         </Container>
       </Navbar>
     );
-  }
 }
 
 Sidebar.defaultProps = {
@@ -268,18 +227,11 @@ Sidebar.defaultProps = {
 };
 
 Sidebar.propTypes = {
-  // links that will be displayed inside the component
   routes: PropTypes.arrayOf(PropTypes.object),
   logo: PropTypes.shape({
-    // innerLink is for links that will direct the user within the app
-    // it will be rendered as <Link to="...">...</Link> tag
     innerLink: PropTypes.string,
-    // outterLink is for links that will direct the user outside the app
-    // it will be rendered as simple <a href="...">...</a> tag
     outterLink: PropTypes.string,
-    // the image src of the logo
     imgSrc: PropTypes.string.isRequired,
-    // the alt for the img
     imgAlt: PropTypes.string.isRequired
   })
 };
