@@ -1,28 +1,9 @@
-/*!
-
-=========================================================
-* Argon Dashboard React - v1.1.0
-=========================================================
-
-* Product Page: https://www.creative-tim.com/product/argon-dashboard-react
-* Copyright 2019 Creative Tim (https://www.creative-tim.com)
-* Licensed under MIT (https://github.com/creativetimofficial/argon-dashboard-react/blob/master/LICENSE.md)
-
-* Coded by Creative Tim
-
-=========================================================
-
-* The above copyright notice and this permission notice shall be included in all copies or substantial portions of the Software.
-
-*/
-/*eslint-disable*/
 import React from "react";
 
-// reactstrap components
+// Components
 import { Row, Col, Nav, NavItem, NavLink } from "reactstrap";
 
-class Footer extends React.Component {
-  render() {
+const AdminFooter = () => {
     return (
       <footer className="footer">
         <Row className="align-items-center justify-content-xl-between">
@@ -66,7 +47,6 @@ class Footer extends React.Component {
         </Row>
       </footer>
     );
-  }
 }
 
-export default Footer;
+export default AdminFooter;
