@@ -1,28 +1,9 @@
-/*!
-
-=========================================================
-* Argon Dashboard React - v1.1.0
-=========================================================
-
-* Product Page: https://www.creative-tim.com/product/argon-dashboard-react
-* Copyright 2019 Creative Tim (https://www.creative-tim.com)
-* Licensed under MIT (https://github.com/creativetimofficial/argon-dashboard-react/blob/master/LICENSE.md)
-
-* Coded by Creative Tim
-
-=========================================================
-
-* The above copyright notice and this permission notice shall be included in all copies or substantial portions of the Software.
-
-*/
-/*eslint-disable*/
 import React from "react";
 
-// reactstrap components
+// Components
 import { NavItem, NavLink, Nav, Container, Row, Col } from "reactstrap";
 
-class Login extends React.Component {
-  render() {
+const AuthFooter = () => {
     return (
       <>
         <footer className="py-5">
@@ -31,13 +12,13 @@ class Login extends React.Component {
               <Col xl="6">
                 <div className="copyright text-center text-xl-left text-muted">
                   © 2022{" "}
-                  <a
+                  <NavLink
                     className="font-weight-bold ml-1"
                     href="https://github.com/samisosa20"
                     target="_blank"
                   >
                     Sammy Guttman
-                  </a>
+                  </NavLink>
                 </div>
               </Col>
               <Col xl="6">
@@ -50,22 +31,6 @@ class Login extends React.Component {
                       Sammy Guttman
                     </NavLink>
                   </NavItem>
-                  {/*<NavItem>
-                    <NavLink
-                      href="https://www.creative-tim.com/presentation?ref=adr-auth-footer"
-                      target="_blank"
-                    >
-                      About Us
-                    </NavLink>
-                  </NavItem>
-                  <NavItem>
-                    <NavLink
-                      href="http://blog.creative-tim.com?ref=adr-auth-footer"
-                      target="_blank"
-                    >
-                      Blog
-                    </NavLink>
-                  </NavItem>*/}
                   <NavItem>
                     <NavLink
                       href="https://github.com/creativetimofficial/argon-dashboard/blob/master/LICENSE.md?ref=adr-auth-footer"
@@ -81,7 +46,6 @@ class Login extends React.Component {
         </footer>
       </>
     );
-  }
 }
 
-export default Login;
+export default AuthFooter;
